# server/go-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "轻小说阅读API服务",
        "title": "Light Novel API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/favorites": {
            "get": {
                "description": "获取用户收藏的小说列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户收藏列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "X-Device-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Novel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/favorites/{novel_id}": {
            "post": {
                "description": "添加小说到收藏列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "添加收藏",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "X-Device-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "小说ID",
                        "name": "novel_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "从收藏列表中移除小说",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "取消收藏",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "X-Device-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "小说ID",
                        "name": "novel_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/favorites/{novel_id}/check": {
            "get": {
                "description": "检查小说是否已收藏",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "检查是否已收藏",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "X-Device-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "小说ID",
                        "name": "novel_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/reading/records": {
            "post": {
                "description": "添加一条阅读记录，同时更新阅读统计和已读章节记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "阅读记录"
                ],
                "summary": "添加阅读记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "device-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "阅读记录信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddReadRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Novel": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "readCount": {
                    "description": "阅读量",
                    "type": "integer"
                },
                "status": {
                    "description": "连载中、已完结",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "volumeCount": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "v1.AddReadRecordRequest": {
            "type": "object",
            "required": [
                "chapterNumber",
                "novelId",
                "readDuration",
                "source",
                "volumeNumber"
            ],
            "properties": {
                "chapterNumber": {
                    "type": "integer",
                    "minimum": 1
                },
                "endPosition": {
                    "type": "integer",
                    "minimum": 0
                },
                "isComplete": {
                    "type": "boolean"
                },
                "novelId": {
                    "type": "string"
                },
                "readDuration": {
                    "type": "integer",
                    "minimum": 1
                },
                "source": {
                    "type": "string",
                    "enum": [
                        "web",
                        "app",
                        "other"
                    ]
                },
                "startPosition": {
                    "type": "integer",
                    "minimum": 0
                },
                "volumeNumber": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
    }
}